Add app information section
This section contains information about the app such as its build
and version numbers.

It also provides links to view the source code, license, and to
submit feedback through YouTrack. The license and feedback pages
open in a custom in-app browser view, while the source code opens
in Safari itself. Whether these should be consistent in one way or
another is to be determined.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>