:bug: Resolve issues with in-app browser
When tapping on a link that opens the in-app browser, the page may
not load the correct link. Because values are set once a sheet is
created, the value was not being updated correctly.

To fix this, a new SettingsSafariWindow was introduced to bind the
AppLink value, so that the changes are updated correctly in real
time.

Additionally, some parts of the code in Settings have been refactored
to use chaining closures.

Source: https://stackoverflow.com/questions/64551580/swiftui-sheet-doesnt-access-the-latest-value-of-state-variables-on-first-appear

Signed-off-by: Marquis Kurt <software@marquiskurt.net>