:sparkles: Add context menus for list items
GeneratorList and DictionaryList now have context menu items on
each row in the list. For Generator, this includes saving the
sniglet or listening to it. For DictionaryList, this includes
listening to the sniglet or deleting it.

Additionally, deleting saved sniglets now does so with an
animation to make the process feel smoother.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>