:construction: Add base Apple Watch app
There exists a new Apple Watch target in the project, which displays
a simple app where users can tap to generate sniglets, or long press
on the sniglet to save it to their dictionary and sync it through
iCloud.

The icon for the Apple Watch app still needs to be created for this,
but this is a good start.

^ABY-2 work Development 30m

Signed-off-by: Marquis Kurt <software@marquiskurt.net>