:lipstick: Update share sheet location on Mac Catalyst
In previous releases with the share sheet, the share sheet would
appear after a popover appeared. While it renders correctly on iOS
and iPadOS, it fails to render correctly on macOS, instead showing
a completely blank screen before showing the
NSShareServicePicker.

To fix this, the code for opening the share sheet has been changed
to check for the Mac Catalyst environment and to display the share
sheet controller manually.