:lipstick: Add toast notification support
Previously, when a user saves a sniglet into their personal
dictionary, an alert would appear, letting them know that
the operation went successfully and that the sniglet can be
found in their dictionary to make the necessary edits.

However, this approach is slightly problematic, because it
is a bit too intrusive. To fix this, a new toast notification was
introduced to show a less intrusive toast at the top of the
screen when a sniglet is saved to the dictionary. Functionally,
it serves the same purpose as the alert; however, users don't
need to act on it like the alert.