:sparkles: Add option to share sniglet as text
For some devices and scenarios, saving an image version of
a saved sniglet doesn't make sense. In those scenarios,
using text might be preferred.

To allow this, a new setting has been introduced to let the
share sheet share a text snippet of the saved sniglet instead
of an image.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>