:fire: Remove mentions of Give Me A Sniglet+
Give Me A Sniglet+ was intended to be an in-app purchase upgrade
to unlock additional functionality to the app such as saved sniglets.
The decision has been made to remove this feature.

In future commits, there may be a possibility of implementing a "buy
me a coffee" style IAPs to support development of the project.