:recycle: Make use of SwiftUI control size in dialog
In the "What's New" dialog, the "Continue" button used a custom
style to implement the big, prominent features. Some of the styles
are unnecessary since, in SwiftUI 3, these can be controlled with
headerProminence and control size. This commit refactors the
button in the "What's New" dialog to utilize these features instead.