:children_crossing: Show toast when copying word
In previous releases, tapping on a word to copy it to the clipboard
with "Tap to Copy" turned on did not provide visual feedback. This
commit resolves this by displaying a toast notification after tapping
the word to copy it. This also necessitated creating a new button
view to handle this: TapToCopyButton.

^ABY-14 stage Review work 5m