:bug: Enforce a max character increase
In previous iterations, the app could possibly crash from creating a
range with the same min and max character number. This causes
a problem because Swift interprets this range as empty, meaning
that a fatal error like in ABY-12 could occur.

This commit attempts to address this by checking the maxChars
bound and adding one to it if it is the same size as minChars. This
should effectively fix ABY-12.

^ABY-12 stage Review work 5m