:construction: Add generator page
The Generator page displays a table of sniglets with their confidence.
When clicking on a sniglet in the table, additional toolbar items appear,
allowing the user to interact with the sniglet. Due to current limitations,
it is not possible to add a context menu to the table rows just yet. This
will be improved in a future commit.

^ABY-11 work Development 1h 20m