:children_crossing: Make improvements to Generator design
A new context menu has been added to rows of the generator table
that allow users to access the same actions as in the toolbar, with
the stipulation that the user click on the text of the row exactly, either
in the Sniglet or the Confidence columns.

A new localizable dictionary for macOS has been added, along with
a String extension to fetch localized strings from that dictionary. This
has been added to add Mac-specific keys to localization.

Additionally, you can now sort the generator table by confidence
scores.

^ABY-11 work Development 15m