:construction: Add Dictionary page to Mac app
This commit introduces the redesigned Dictionary section of the app
on macOS. The dictionary uses a table, like the generator, to display
sniglets, confidence scores, and any notes that the user has added.
Upon selecting an entry in the table, the user can perform actions
in the toolbar as necessary, short of editing the note attached. This
functionality will be added in a future commit.

A new SharingServicePicker has been added to the Mac variant,
mainly taking inspiration from a StackOverflow article on how to
bridge the NSSharingServicePicker from AppKit over to SwiftUI.
Some minor adjustments have been made to make the picker align
with the share button better.

A new shareableText extension method on both SavedWord and
Sniglet.Result has been added to generate the share text introduced
in v1.1 of Give Me A Sniglet. On the Mac side, only the text version
will be shared instead of generating an entire image.

In the Generator table, selected sniglets can now be shared with the
SharingServicePicker. Additionally, to add visual separation between
selected sniglet actions and the overall view, a spacer has been added
between the two toolbar groups.

^ABY-11 work Development 2h