♻️ Refactor sharing system for iOS (ABY-17)
This commit introduces some changes to the sharing system on the iOS
app. The views that show a sniglet and its associated data now conform
to the SnigletShareable protocol, which allows that view to show a share
sheet with some data.

This commit also changes the structure of some toolbar to show a menu
when necessary. This may change with iOS/iPadOS 16 with support for
desktop-class iPad apps.

^ABY-17 stage Review work Development 1h