:construction: Change generator table view for iOS
To be more in line with the updated Mac version of the app, the new
iPad layout will utilize the Table view instead of the typical
NavigationSplitView that the app was using before. This allows for
more desktop like interactions, making this a desktop-class app.

However, this new view does NOT support the sharing system yet, as
this is still being worked on. Likewise, the Dictionary view does
not support this new interface yet.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>